requirements.txt
Flask>=2.2
Flask-Cors>=3.0
Flask-SQLAlchemy>=3.1
gunicorn>=20.1
psutil>=5.9
requests>=2.31
Werkzeug>=2.2
PyJWT>=2.8
